+++ src/bl64-aws-setup.bash
@@ -38,7 +38,11 @@ function bl64_aws_setup() {
   bl64_dbg_lib_show_function "$@"
   local aws_bin="${1:-${BL64_VAR_DEFAULT}}"
 
-  _bl64_aws_set_command "$aws_bin" &&
+  bl64_check_module_imported 'BL64_CHECK_MODULE' &&
+    bl64_check_module_imported 'BL64_DBG_MODULE' &&
+    bl64_check_module_imported 'BL64_MSG_MODULE' &&
+    bl64_check_module_imported 'BL64_FS_MODULE' &&
+    _bl64_aws_set_command "$aws_bin" &&
     _bl64_aws_set_options &&
     _bl64_aws_set_resources &&
     bl64_check_command "$BL64_AWS_CMD_AWS" &&