+++ src/bl64-fs-setup.bash
@@ -22,7 +22,11 @@ function bl64_fs_setup() {
     return 21
   bl64_dbg_lib_show_function
 
-  _bl64_fs_set_command &&
+  bl64_check_module_imported 'BL64_CHECK_MODULE' &&
+    bl64_check_module_imported 'BL64_DBG_MODULE' &&
+    bl64_check_module_imported 'BL64_MSG_MODULE' &&
+    bl64_check_module_imported 'BL64_FMT_MODULE' &&
+    _bl64_fs_set_command &&
     _bl64_fs_set_alias &&
     _bl64_fs_set_options &&
     BL64_FS_MODULE="$BL64_VAR_ON"