+++ src/bl64-tf-env.bash
@@ -2,7 +2,7 @@
 # BashLib64 / Module / Globals / Interact with Terraform
 #######################################
 
-export BL64_TF_VERSION='1.3.0'
+export BL64_TF_VERSION='1.4.0'
 
 # Optional module. Not enabled by default
 export BL64_TF_MODULE="$BL64_VAR_OFF"