+++ src/bl64-tm-setup.bash
@@ -22,7 +22,9 @@ function bl64_tm_setup() {
     return 21
   bl64_dbg_lib_show_function
 
-  BL64_TM_MODULE="$BL64_VAR_ON"
+  bl64_check_module_imported 'BL64_CHECK_MODULE' &&
+    bl64_check_module_imported 'BL64_DBG_MODULE' &&
+    BL64_TM_MODULE="$BL64_VAR_ON"
 
   bl64_check_alert_module_setup 'tm'
 }